# src/lib/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Workers Webhooks API",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Bearer token authentication"
      },
      "apiKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "apiKeyCookie",
        "description": "API Key authentication via cookie"
      }
    },
    "responses": {
      "ErrorResponse": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "description": "Error message"
                }
              }
            },
            "example": {
              "error": "Internal server error"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyCookie": []
    }
  ],
...
  "paths": {
...
    "/endpoint-groups": {
      "get": {
        "summary": "List Endpoint Groups",
        "description": "List endpoint groups for the current environment",
        "tags": [
          "Endpoint Groups"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "endpointGroups": {
                      "type": "array",
                      "description": "List of endpoint groups",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "description": "Endpoint group ID"
                          },
                          "environmentId": {
                            "type": "string",
                            "description": "Environment ID"
                          },
                          "name": {
                            "type": "string",
                            "description": "Name"
                          },
                          "description": {
                            "type": "string",
                            "description": "Description"
                          },
                          "endpointIds": {
                            "type": "array",
                            "description": "List of endpoint IDs"
                          },
                          "enabled": {
                            "type": "boolean",
                            "description": "Whether the endpoint group is enabled"
                          },
                          "createdAt": {
                            "type": "string",
                            "description": "Created at"
                          },
                          "updatedAt": {
                            "type": "string",
                            "description": "Updated at"
                          }
                        }
                      }
                    }
                  },
                  "example": {
                    "endpointGroups": [
                      {
                        "id": "grp_a1b2_efgh5678",
                        "environmentId": "a1b2",
                        "name": "My Endpoint Group",
                        "description": "My Endpoint Group description",
                        "endpointIds": [
                          "ep_a1b2_efgh5678"
                        ],
                        "enabled": true,
                        "createdAt": "2021-01-01T00:00:00.000Z",
                        "updatedAt": "2021-01-01T00:00:00.000Z"
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. Due to missing or invalid authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ErrorResponse",
                  "example": {
                    "error": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. You do not have permission to access this resource or to perform this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ErrorResponse",
                  "example": {
                    "error": "Forbidden"
                  }
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests. You have exceeded the rate limit. Try again later.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ErrorResponse",
                  "example": {
                    "error": "Rate limit exceeded"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error. This is a problem with the server that you cannot fix.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ErrorResponse",
                  "example": {
                    "error": "Internal server error"
                  }
                }
              }
            }
          }
        },
        "x-speakeasy-group": "endpointGroups",
        "x-speakeasy-name-override": "list"
      },
      "post": {
        "summary": "Create Endpoint Group",
        "description": "Create new endpoint group",
        "tags": [
          "Endpoint Groups"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Name of the endpoint group",
                    "example": "My Endpoint Group"
                  },
                  "description": {
                    "type": "string",
                    "description": "Description of the endpoint group",
                    "example": "My Endpoint Group description"
                  },
                  "endpointIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "List of endpoint IDs",
                    "example": [
                      "ep_a1b2_efgh5678"
                    ]
                  },
                  "enabled": {
                    "type": "boolean",
                    "description": "Whether the endpoint group is enabled",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "Endpoint group ID"
                    },
                    "environmentId": {
                      "type": "string",
                      "description": "Environment ID"
                    },
                    "name": {
                      "type": "string",
                      "description": "Name"
                    },
                    "description": {
                      "type": "string",
                      "description": "Description"
                    },
                    "endpointIds": {
                      "type": "array",
                      "description": "List of endpoint IDs",
                      "items": {
                        "type": "string"
                      }
                    },
                    "enabled": {
                      "type": "boolean",
                      "description": "Whether the endpoint group is enabled"
                    },
                    "createdAt": {
                      "type": "string",
                      "description": "Created at"
                    },
                    "updatedAt": {
                      "type": "string",
                      "description": "Updated at"
                    }
                  },
                  "example": {
                    "id": "grp_a1b2_efgh5678",
                    "environmentId": "a1b2",
                    "name": "My Endpoint Group",
                    "description": "My Endpoint Group description",
                    "endpointIds": [
                      "ep_a1b2_efgh5678"
                    ],
                    "enabled": true,
                    "createdAt": "2021-01-01T00:00:00.000Z",
                    "updatedAt": "2021-01-01T00:00:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request. Usually due to missing parameters, or invalid parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ErrorResponse",
                  "example": {
                    "error": "Name is required"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. Due to missing or invalid authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ErrorResponse",
                  "example": {
                    "error": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. You do not have permission to access this resource or to perform this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ErrorResponse",
                  "example": {
                    "error": "Forbidden"
                  }
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests. You have exceeded the rate limit. Try again later.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ErrorResponse",
                  "example": {
                    "error": "Rate limit exceeded"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error. This is a problem with the server that you cannot fix.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ErrorResponse",
                  "example": {
                    "error": "Internal server error"
                  }
                }
              }
            }
          }
        },
        "x-speakeasy-group": "endpointGroups",
        "x-speakeasy-name-override": "create"
      }
    },
...
  }
}